Clearer function names, more comments, mass deletions (see extended description)
All of the competition specific export functions have been deleted and replaced with a more general function to remove confusion and shorten the file. Now on, to export a specific competition, simply fill in the comp and name arguments. If the data requested needs to be specific, simply use the base CSVExport() function.